PhxLocalizedRoutes behaviour (Phoenix Localized Routes v0.1.0)

Macro to create and validate PhxLocalizedRoutes configuration module with convenience callbacks to fetch specific values. It also creates a LiveHelper module to be used in LiveView projects. For maximum performance, most functions return values including additional data computed at compile time.

Types

opts()

@type opts() :: [scopes: scopes_map(), gettext_module: module()]

scope_map()

@type scope_map() :: %{
  assign: %{required(atom()) => any()} | nil,
  scopes: scopes_map() | nil
}

Type that represents a config module map containing scope key/values.

scopes()

@type scopes() :: PhxLocalizedRoutes.Config.scopes()

scopes_map()

@type scopes_map() :: %{required(binary()) => scope_map()}

scopes_nested()

@type scopes_nested() :: %{required(binary()) => PhxLocalizedRoutes.Scope.Nested.t()}

Callbacks

assigned_values(key_or_keys)

@callback assigned_values(key_or_keys :: atom() | String.t() | list()) :: list()

Return a list of unique values assigned to given key. Returns a list of tuples with unique combinations when a list of keys is given.

Example

iex> ExampleWeb.LocalizedRoutes.assigned_values(:locale)
["en", "nl"]

iex> ExampleWeb.LocalizedRoutes.assigned_values([:locale, :locale])
[{"en", "en-GB"}, {"nl", "nl-NL"}, {"nl", "nl_BE"}]

config()

@callback config() :: map()

Returns the configuration with precomputed values and flattened scopes

get_scope(scope_helper)

@callback get_scope(scope_helper :: nil | String.t()) :: map()

Return the configuration of given scope helper

scopes()

@callback scopes() :: map()

Return the scopes in a flat structure

scopes_nested()

@callback scopes_nested() :: map()

Return the scopes in a nested structure