PhxLocalizedRoutes behaviour (Phoenix Localized Routes v0.1.0)

Macro to create and validate PhxLocalizedRoutes configuration module with convenience callbacks to fetch specific values. It also creates a LiveHelper module to be used in LiveView projects. For maximum performance, most functions return values including additional data computed at compile time.

Types

opts()

@type opts() :: [scopes: opts_scopes(), gettext_module: module()]

opts_scope_map()

@type opts_scope_map() :: %{
  assign: %{required(atom()) => any()} | nil,
  scopes: opts_scopes() | nil
}

opts_scopes()

@type opts_scopes() :: %{required(binary()) => opts_scope_map()}

scope()

@type scope() :: PhxLocalizedRoutes.Scope.Flat.t()

scope_nested()

@type scope_nested() :: PhxLocalizedRoutes.Scope.Nested.t()

scopes()

@type scopes() :: PhxLocalizedRoutes.Scope.Flat.kv_map()

scopes_nested()

@type scopes_nested() :: PhxLocalizedRoutes.Scope.Nested.kv_map()

Callbacks

assigned_values(key_or_keys)

@callback assigned_values(key_or_keys :: atom() | String.t() | list()) :: list()

Return a list of unique values assigned to given key. Returns a list of tuples with unique combinations when a list of keys is given.

Example

iex> ExampleWeb.LocalizedRoutes.assigned_values(:locale)
["en", "nl"]

iex> ExampleWeb.LocalizedRoutes.assigned_values([:locale, :locale])
[{"en", "en-GB"}, {"nl", "nl-NL"}, {"nl", "nl_BE"}]

config()

@callback config() :: map()

Returns the configuration with precomputed values and flattened scopes

get_scope(scope_helper)

@callback get_scope(scope_helper :: nil | String.t()) :: scope()

Return the configuration of given scope helper

scopes()

@callback scopes() :: scopes()

Return the scopes in a flat structure

scopes_nested()

@callback scopes_nested() :: scopes_nested()

Return the scopes in a nested structure